fix: bpm 为小数导致歌在 mai 里加载不了

+++ Sitreamai/Models/MusicXml.cs
@@ -336,7 +336,7 @@ public class MusicXml
         set
         {
             Modified = true;
-            RootNode.SelectSingleNode("bpm").InnerText = value.ToString();
+            RootNode.SelectSingleNode("bpm").InnerText = ((int)value).ToString();
         }
     }
